fix(pr-review): extract DatalogPrettyPrinter from cli.py into dedicated module
Addresses PR review comment about DatalogPrettyPrinter being too large for cli.py. Moved to neurolang/frontend/datalog/pretty_printer.py alongside the SQUALL-related code.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
